+++ src/main/java/com/uber/rss/messages/TaskAttemptCommitStateItem.java
@@ -49,7 +49,7 @@ public class TaskAttemptCommitStateItem extends BaseMessage {
         buf.writeInt(appShuffleId.getShuffleId());
         buf.writeInt(mapTaskAttemptIds.size());
         for (MapTaskAttemptId entry: mapTaskAttemptIds) {
-            buf.writeLong(entry.getMapId());
+            buf.writeInt(entry.getMapId());
             buf.writeLong(entry.getTaskAttemptId());
         }
         buf.writeInt(partitionFilePathAndLengths.size());
@@ -67,7 +67,7 @@ public class TaskAttemptCommitStateItem extends BaseMessage {
         int count = buf.readInt();
         List<MapTaskAttemptId> mapTaskAttemptIdList = new ArrayList<>(count);
         for (int i = 0; i < count; i++) {
-            long mapId = buf.readLong();
+            int mapId = buf.readInt();
             long taskAttemptId = buf.readLong();
             mapTaskAttemptIdList.add(new MapTaskAttemptId(mapId, taskAttemptId));
         }
